+++ control/src/cache/reconcile/taper.rs
@@ -44,6 +44,8 @@ pub fn log(position: f32) -> f32 {
     value + delta_to_next * remainder
 }
 
+// TODO
+#[allow(unused)]
 pub fn reverse_log(position: f32) -> f32 {
     1.0 - log(1.0 - position)
 }