+++ control/src/cache/reconcile/tone.rs
@@ -6,7 +6,7 @@ impl Store {
     pub fn reconcile_tone(&mut self) {
         self.cache.attributes.tone = calculate(
             self.input.tone.value(),
-            self.control_for_attribute(AttributeIdentifier::Tone),
+            self.control_value_for_attribute(AttributeIdentifier::Tone),
             (0.0, 1.0),
             None,
         );