+++ control/src/input/button.rs
@@ -1,7 +1,5 @@
 //! Manage button's state.
 
-use super::clock::ClockDetector;
-
 /// Use this to hold buttons state over time.
 ///
 /// Detects clicking, holding, or tapped-in tempo.
@@ -11,7 +9,6 @@ pub struct Button {
     pub pressed: bool,
     pub clicked: bool,
     pub held: u32,
-    clock_detector: ClockDetector,
 }
 
 impl Button {
@@ -24,15 +21,6 @@ impl Button {
         } else {
             0
         };
-
-        self.clock_detector.tick();
-        if self.clicked {
-            self.clock_detector.trigger();
-        }
-    }
-
-    pub fn detected_tempo(&self) -> Option<u32> {
-        self.clock_detector.tempo
     }
 }
 
@@ -66,67 +54,5 @@ mod tests {
         assert_eq!(button.held, 3);
         button.update(false);
         assert_eq!(button.held, 0);
-    }
-
-    #[test]
-    fn when_clicked_in_exact_interval_it_detects_tempo() {
-        let mut button = Button::default();
-        for _ in 0..4 {
-            for _ in 0..1999 {
-                button.update(false);
-            }
-            button.update(true);
-        }
-        assert_eq!(button.clock_detector.tempo, Some(2000));
-    }
-
-    #[test]
-    fn when_clicked_in_rough_interval_within_toleration_it_detects_tempo() {
-        let mut button = Button::default();
-        button.update(true);
-        for _ in 0..1990 {
-            button.update(false);
-        }
-        button.update(true);
-        for _ in 0..2059 {
-            button.update(false);
-        }
-        button.update(true);
-        for _ in 0..1999 {
-            button.update(false);
-        }
-        button.update(true);
-        assert_eq!(button.clock_detector.tempo, Some(2000));
-    }
-
-    #[test]
-    fn when_clicked_too_fast_it_does_not_detect_tempo() {
-        let mut button = Button::default();
-        for _ in 0..4 {
-            for _ in 0..19 {
-                button.update(false);
-            }
-            button.update(true);
-        }
-        assert_eq!(button.clock_detector.tempo, None);
-    }
-
-    #[test]
-    fn when_clicked_in_unequal_interval_it_does_not_detect_tempo() {
-        let mut button = Button::default();
-        button.update(true);
-        for _ in 0..1999 {
-            button.update(false);
-        }
-        button.update(true);
-        for _ in 0..1999 {
-            button.update(false);
-        }
-        button.update(true);
-        for _ in 0..1083 {
-            button.update(false);
-        }
-        button.update(true);
-        assert_eq!(button.clock_detector.tempo, None);
     }
 }